requirements.txt
deep-lynx>=0.1.7
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
requests>=2.31.0
urllib3>=2.0.0 